fake session handling in dev+test
idea:
in dev & test env this implementation is used instead of th real one (which is used in prod)

must behave like the real version!
the `authenticated-entity` object must have the same shape as in prod, so should look like `user_session` not `user`. likely for now a simple object is enough `{:user_id "abcdef"}